refactor(bridge): drop dead request-client injection and finish restricted-file consolidation
- Replace the injected `final` AbortableRequestClient (never substitutable
  outside its library) with a top-level sendAbortableRequest; the http.Client
  remains the only transport seam.
- Fold saveTokens into writeRestrictedFile as Step 18 planned, so the token,
  bridge-id, and onboarding-marker files share one owner-only writer.
- Replace the unused local 401 HTTP server in bridge_runtime_auth_test with a
  401 MockClient repository, restoring the refresh-rejected startup coverage.
- Remove an unreachable isAborted re-check and rename tokenManager locals.
