+++ bridge/app/test/auth/auth_api_registration_test.dart
@@ -3,7 +3,6 @@ import "dart:io";
 
 import "package:http/http.dart" as http;
 import "package:sesori_bridge/src/auth/auth_api.dart";
-import "package:sesori_bridge/src/foundation/abortable_request_client.dart";
 import "package:test/test.dart";
 
 void main() {
@@ -108,7 +107,6 @@ AuthApi _createApi(_BridgesTestServer server) {
   return AuthApi(
     authBackendUrl: server.baseUrl,
     client: client,
-    requestClient: const AbortableRequestClient(),
     requestDeadline: AuthApi.defaultRequestDeadline,
   );
 }
