+++ bridge/app/test/auth/login_test.dart
@@ -10,7 +10,6 @@ import 'package:sesori_bridge/src/auth/auth_api.dart';
 import 'package:sesori_bridge/src/auth/login_email_api.dart';
 import 'package:sesori_bridge/src/auth/login_email_repository.dart';
 import 'package:sesori_bridge/src/auth/login_oauth_service.dart';
-import 'package:sesori_bridge/src/foundation/abortable_request_client.dart';
 import 'package:sesori_shared/sesori_shared.dart';
 import 'package:test/test.dart';
 
@@ -279,7 +278,6 @@ void main() {
         final api = AuthApi(
           authBackendUrl: authServer.baseUrl,
           client: authServer.client,
-          requestClient: const AbortableRequestClient(),
           requestDeadline: AuthApi.defaultRequestDeadline,
         );
 
@@ -530,7 +528,6 @@ LoginOAuthService _createOAuthService({
     api: AuthApi(
       authBackendUrl: authServer.baseUrl,
       client: authServer.client,
-      requestClient: const AbortableRequestClient(),
       requestDeadline: AuthApi.defaultRequestDeadline,
     ),
     clientType: AuthClientType.bridgeMacos,
