+++ bridge/app/test/auth/profile_test.dart
@@ -4,7 +4,6 @@ import 'dart:io';
 import 'package:http/http.dart' as http;
 import 'package:sesori_bridge/src/auth/auth_api.dart';
 import 'package:sesori_bridge/src/auth/auth_repository.dart';
-import 'package:sesori_bridge/src/foundation/abortable_request_client.dart';
 import 'package:test/test.dart';
 
 void main() {
@@ -20,7 +19,6 @@ void main() {
         api: AuthApi(
           authBackendUrl: 'http://${server.address.host}:${server.port}',
           client: client,
-          requestClient: const AbortableRequestClient(),
           requestDeadline: AuthApi.defaultRequestDeadline,
         ),
       );
