+++ bridge/app/test/auth/validate_test.dart
@@ -4,7 +4,6 @@ import 'dart:io';
 
 import 'package:http/http.dart' as http;
 import 'package:sesori_bridge/src/auth/auth_api.dart';
-import 'package:sesori_bridge/src/foundation/abortable_request_client.dart';
 import 'package:test/test.dart';
 
 void main() {
@@ -19,7 +18,6 @@ void main() {
       api = AuthApi(
         authBackendUrl: 'http://${server.address.host}:${server.port}',
         client: client,
-        requestClient: const AbortableRequestClient(),
         requestDeadline: AuthApi.defaultRequestDeadline,
       );
     });
@@ -143,7 +141,6 @@ void main() {
         final deadlineApi = AuthApi(
           authBackendUrl: 'https://auth.example.test',
           client: abortAwareClient,
-          requestClient: const AbortableRequestClient(),
           requestDeadline: Duration.zero,
         );
