+++ bridge/app/test/repositories/app_client_status_repository_test.dart
@@ -1,7 +1,7 @@
 import "package:sesori_bridge/src/api/models/app_client_status_response.dart";
 import "package:sesori_bridge/src/api/models/generate_session_metadata_response.dart";
 import "package:sesori_bridge/src/api/sesori_server_api.dart";
-import "package:sesori_bridge/src/foundation/abortable_request_client.dart";
+import "package:sesori_bridge/src/foundation/abortable_request.dart";
 import "package:sesori_bridge/src/repositories/app_client_status_repository.dart";
 import "package:sesori_shared/sesori_shared.dart";
 import "package:test/test.dart";
